testing mult trans

+++ core/views.py
@@ -2227,29 +2227,34 @@ def test_transformation_scenario(request):
 
 		# get record
 		record = models.Record.objects.get(id=request.POST.get('db_id'))
+		record_iter = models.Record.objects.get(id=request.POST.get('db_id'))		
 
 		try:
 
+			# testing multiple, chained transformations
 			if request.POST.get('trans_test_type') == 'multiple':
 				
-				########################################################################################################
-				# MULTIPLE
-				########################################################################################################
+				# get and rehydrate sel_trans_json
+				sel_trans = json.loads(request.POST.get('sel_trans_json'))				
 
-				'''
-				  - loop through passed sel_trans_json
-					  - init Transformation
-					  - transform same record instance, never saving, but modifying
-				'''
-				
-				########################################################################################################
-				pass
+				# loop through transformations
+				for trans in sel_trans:
 
+					# init Transformation instance
+					trans = models.Transformation.objects.get(pk=int(trans['trans_id']))
+
+					# transform with record
+					trans_results = trans.transform_record(record_iter)
+
+					# set to record.document for next iteration
+					record_iter.document = trans_results
+
+				# finally, fall in line with trans_results as record_iter document string
+				trans_results = record_iter.document
+
+			# testing single transformation
 			elif request.POST.get('trans_test_type') == 'single':
 				
-				########################################################################################################
-				# SINGLE
-				########################################################################################################
 				# init new transformation scenario
 				trans = models.Transformation(
 					name='temp_trans_%s' % str(uuid.uuid4()),
@@ -2258,13 +2263,11 @@ def test_transformation_scenario(request):
 				)
 				trans.save()
 
-
-				# validate with record
-				trans_results = trans.transform_record(record)			
+				# transform with record
+				trans_results = trans.transform_record(record)
 
 				# delete temporary trans
 				trans.delete()
-				########################################################################################################
 
 			# if raw transformation results
 			if response_type == 'transformed_doc':
@@ -2297,10 +2300,12 @@ def test_transformation_scenario(request):
 				return HttpResponse(diff_html, content_type="text/xml")
 			
 		except Exception as e:
-
 			logger.debug('test validation scenario was unsucessful, deleting temporary vs')
-			trans.delete()
-
+			try:
+				if request.POST.get('trans_test_type') == 'single':
+					trans.delete()
+			except:
+				logger.debug('could not delete temporary transformation')
 			return HttpResponse(str(e), content_type="text/plain")
